requirements/base.txt
asgiref==3.8.1
Django==5.0.6
django-debug-toolbar==4.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
tzdata==2024.1
